Refactor InspectorTable to be configurable
