+++ experimental/PropertyDDS/packages/property-inspector-table/package.json
@@ -45,7 +45,7 @@
     "classnames": "^2.2.6",
     "lodash.debounce": "^4.0.8",
     "memoize-one": "^5.0.0",
-    "react-base-table": "1.0.2",
+    "react-base-table": "1.13.2",
     "react-loading-skeleton": "^1.1.2",
     "react-select": "^2.4.3",
     "react-virtualized-auto-sizer": "^1.0.2"
@@ -66,6 +66,7 @@
     "@storybook/addon-essentials": "^6.4.22",
     "@storybook/addon-links": "^6.4.22",
     "@storybook/react": "^6.4.22",
+    "@types/enzyme": "3.10.12",
     "@types/jest": "22.2.3",
     "@types/react": "^16.9.15",
     "@types/underscore": "^1.9.1",
