+++ experimental/PropertyDDS/packages/property-inspector-table/src/InspectorTableTypes.ts
@@ -5,8 +5,77 @@
 
 import { BaseProxifiedProperty } from "@fluid-experimental/property-proxy";
 import { BaseProperty } from "@fluid-experimental/property-properties";
+
+import { BaseTableProps, SortOrder } from "react-base-table";
 import { IRepoExpiryGetter, IRepoExpirySetter } from "./CommonTypes";
-import { IInspectorSearchState } from "./utils";
+
+export interface SearchResult {
+  foundMatches?: IInspectorSearchMatch[];
+  matchesMap?: IInspectorSearchMatchMap;
+  currentResult?: number;
+}
+export interface IShowNextResultResult {
+  /**
+   * New rows needed to expand to show next results
+   */
+  expandedRows: IExpandedMap;
+  /**
+   * row index of the next result. This value is used to automatically scroll in the table
+   */
+  rowIdx: number;
+  /**
+   * Next result column index
+   */
+  columnIdx: number;
+}
+
+interface ISearchLevelState {
+  data?: IInspectorRow[];
+  index: number;
+}
+
+export interface IInspectorSearchState {
+  abort?: boolean;
+  newMatchFound?: boolean;
+  updateScheduled?: number;
+  chunkSize?: number;
+  expression?: string;
+  foundMatches: IInspectorSearchMatch[];
+  levels?: ISearchLevelState[];
+  matchesMap: IInspectorSearchMatchMap;
+  scheduled?: number;
+  childToParentMap: { [key: string]: string; };
+}
+
+export interface IInspectorSearchControls {
+  abortHandler: IInspectorSearchAbortHandler;
+  state: IInspectorSearchState;
+}
+export interface IRowData<T = never> {
+	data?: T;
+	id: string;
+	children?: IRowData<T>[];
+	// TODO revisit those types which required for propertyDDS
+	isReference?: boolean;
+	context?: string;
+  isNewDataRow?: boolean;
+}
+
+export type IToTableRowsProps = Pick<IInspectorTableProps,
+  "dataCreationHandler" | "dataCreationOptionGenerationHandler" | "childGetter" | "nameGetter" | "readOnly">;
+
+export interface IToTableRowsOptions {
+  depth: number;
+  addDummy: boolean;
+  followReferences: boolean;
+  ascending: boolean;
+  parentIsConstant?: boolean;
+}
+
+export interface IPropertyToTableRowOptions extends Partial<IToTableRowsOptions> {
+  depth: number;
+  dataCreation: boolean;
+}
 
 export interface IColumns {
   dataGetter?: (params: IDataGetterParameter) => React.ReactNode | null;
@@ -22,12 +91,10 @@ export interface IColumns {
 /**
  * The interface for an entry of the visualization data array.
  */
-export interface IInspectorRow {
+export interface IInspectorRow extends IRowData<BaseProxifiedProperty>{
   context: string;
-  data: any;
-  children: IInspectorRow[] | undefined;
-  id: string;
   isConstant: boolean;
+  children?: IInspectorRow[];
   isReference: boolean;
   name: string;
   typeid: string;
@@ -59,9 +126,11 @@ export interface IDataCreationOptions {
   options?: any;
 }
 
-export type IInspectorColumnsKeys = "context" | "name" | "type" | "value";
+export type IInspectorColumnsKeys = "name" | "type" | "value";
 
-export interface IInspectorTableProps {
+type BaseTablePropsPartial<T> = Omit<BaseTableProps<T>, "columns"> ;
+
+export interface IInspectorTableProps<T extends IRowData<T> = any> extends BaseTablePropsPartial<T> {
   /**
    * The active repository guid. Used as the prefix for the table row ids.
    */
@@ -74,7 +143,7 @@ export interface IInspectorTableProps {
    * A callback to override the child of a row,
    * defaults to the hierarchical child of the property the row visualizes.
    */
-  childGetter?: (child: any, name: string, parent: BaseProperty, typeid: string, context: string) => any;
+  childGetter?: (child: any, name: string, parent: any, typeid: string, context: string) => any;
   /**
    * Current Urn
    */
@@ -86,16 +155,20 @@ export interface IInspectorTableProps {
   /**
    * The raw data to be visualized.
    */
-  data?: BaseProxifiedProperty;
+  data?: any;
+  /**
+   *  Reference property handler
+   */
+  editReferenceHandler?: (rowData: T, newPath: string) => Promise<void>;
   /**
    * A callback function that is called on data creation. If not specified,
    * data creation will be disabled.
    */
-  dataCreationHandler?: (rowData: IInspectorRow, name: string, typeid: string, context: string) => Promise<any>;
+  dataCreationHandler?: (rowData: T, name: string, typeid: string, context: string) => Promise<any>;
   /**
    * A callback that is executed to compute the name and the options available for the data creation.
    */
-  dataCreationOptionGenerationHandler?: (rowData: IInspectorRow, nameOnly: boolean) => IDataCreationOptions;
+  dataCreationOptionGenerationHandler?: (rowData: T, nameOnly: boolean) => IDataCreationOptions;
   /**
    * A handler to delete (expire) a repository.
    */
@@ -111,7 +184,7 @@ export interface IInspectorTableProps {
   /**
    * A callback to override the name of a row, defaults to the id of the property the row visualizes.
    */
-  nameGetter?: (name: string, parent: BaseProperty, typeid: string, context: string) => string;
+  nameGetter?: (name: string, parent: any, typeid: string, context: string) => string;
   /**
    * Indicates if the table is in read only mode
    */
@@ -119,7 +192,7 @@ export interface IInspectorTableProps {
   /**
    * Callback that is invoked to determine the icon.
    */
-  rowIconRenderer?: (rowData: IInspectorRow) => React.ReactNode;
+  rowIconRenderer?: (rowData: T) => React.ReactNode;
   /**
    * Width of the table
    */
@@ -160,6 +233,12 @@ export interface IInspectorTableProps {
    * if checkout is in progress
    */
   checkoutInProgress: boolean;
+  /**
+   * A transformation function from raw data to table rows
+   */
+  toTableRows: (row: any, props: IToTableRowsProps,
+    options?: Partial<IToTableRowsOptions>, pathPrefix?: string) => T[];
+  columnsRenderers?: Record<string, (...props: any) => any>;
 }
 
 export interface IInspectorSearchMatch {
@@ -175,7 +254,7 @@ export interface IInspectorSearchMatch {
 }
 
 export type IInspectorSearchCallback = (foundMatches: IInspectorSearchMatch[], matchesMap: IInspectorSearchMatchMap,
-                                        done: boolean, childToParentMap: { [key: string]: string; }) => void;
+  done: boolean, childToParentMap: { [key: string]: string; }) => void;
 
 export type IInspectorSearchAbortHandler = () => void;
 
@@ -187,13 +266,13 @@ export interface IExpandedMap {
   [key: string]: boolean;
 }
 
-export interface IInspectorTableState {
+export interface IInspectorTableState<T = any> {
   childToParentMap: { [key: string]: string; };
-  commitHistoryVisible: boolean;
+  commitHistoryVisible?: boolean;
   currentResult?: number;
   foundMatches: IInspectorSearchMatch[];
   matchesMap: IInspectorSearchMatchMap;
-  editReferenceRowData: IInspectorRow | null;
+  editReferenceRowData: T | null;
   expanded: IExpandedMap;
   expandedRepoGuid: string | undefined;
   expandedRepoMap: { [key: string]: IExpandedMap; };
@@ -203,6 +282,42 @@ export interface IInspectorTableState {
   searchInProgress: boolean;
   searchState?: IInspectorSearchState;
   showFormRowID: string;
-  sortBy: { [key: string]: string; };
-  tableRows: IInspectorRow[];
+  sortBy: {
+    key: React.Key;
+    order: SortOrder;
+  };
+  tableRows: T[];
+}
+
+export interface ColumnRendererType {
+  rowData: IInspectorRow;
+  cellData: React.ReactNode | undefined;
+  columnIndex: number;
+  tableProps: IInspectorTableProps;
+  searchResult: SearchResult;
+  renderCreationRow: any;
+  referenceHandler: {
+    initialReferenceEdit: any;
+    cancelReferenceEdit: any;
+  };
+}
+
+export interface IEditableValueCellProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  /**
+   * Indicates whether we are following references or not.
+   * Defaults to true.
+   */
+  followReferences?: boolean;
+  /**
+   * A callback that returns the icons based on the row data.
+   */
+  iconRenderer: (rowData: IInspectorRow) => React.ReactNode;
+  /**
+   * The row data of the row which contains the cell.
+   */
+  rowData: IInspectorRow;
+  /**
+   * Indicates if read only mode is enabled
+   */
+  readOnly: boolean;
 }
