+++ experimental/PropertyDDS/packages/property-inspector-table/src/NewDataRow.tsx
@@ -27,17 +27,13 @@ export interface INewDataRowProps {
    * The data type to be added: Asset, Component or Property
    */
   dataType: string;
-  /**
-   * Callback that is invoked if the row is clicked.
-   */
-  onClick: () => void;
 }
 
 class NewDataRow extends React.Component<INewDataRowProps & WithStyles<typeof styles>> {
   public render() {
     const { classes } = this.props;
     return (
-      <div className={classNames(classes.row)} onClick={this.props.onClick}>
+      <div className={classNames(classes.row)}>
         <SvgIcon
           svgId={"plus-24"}
           height={iconHeight}
