+++ experimental/PropertyDDS/packages/property-inspector-table/src/TypeColumn.tsx
@@ -105,3 +105,5 @@ export const TypeColumn: React.FunctionComponent<ITypeColumn> = ({ rowData }) =>
     </Tooltip>
   );
 };
+
+export const useChipStyles = useStyles;
