+++ experimental/PropertyDDS/packages/property-inspector-table/src/index.ts
@@ -4,8 +4,12 @@
  */
 
 export * from "./InspectorTable";
-export * from "./InspectorTableTypes";
 export * from "./PropertyDataCreationHandlers";
+export * from "./propertyInspectorUtils";
 export * from "./icons";
 export * from "./ModalManager";
 export * from "./ModalRoot";
+export * from "./CustomChip";
+export * from "./TypeColumn";
+export * from "./utils";
+export * from "./InspectorTableTypes";
