+++ experimental/PropertyDDS/packages/property-inspector-table/src/webpack.index.ts
@@ -9,8 +9,12 @@ import InspectorTableIcons from "../assets/icons/SVGStoreIcons";
 export { InspectorTableIcons };
 
 export * from "./InspectorTable";
-export * from "./InspectorTableTypes";
 export * from "./PropertyDataCreationHandlers";
+export * from "./propertyInspectorUtils";
 export * from "./icons";
 export * from "./ModalManager";
 export * from "./ModalRoot";
+export * from "./CustomChip";
+export * from "./TypeColumn";
+export * from "./utils";
+export * from "./InspectorTableTypes";
